requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
mcp==1.0.0
anyio==3.7.1
bcrypt==4.1.2
